separate AdminPrepareEmailWorkerAction to two action classes, one each for address mode and group mode
This is another case of wrong abstraction. Both actions have close to nothing in common, other than “preparing admin email” part.
