+++ src/main/webapp/WEB-INF/queue.xml
@@ -7,7 +7,16 @@
 </queue>
 
     <queue>
-        <name>admin-prepare-email-queue</name>
+        <name>admin-prepare-email-address-mode-queue</name>
+        <rate>1/s</rate>
+        <bucket-size>1</bucket-size>
+        <retry-parameters>
+            <task-retry-limit>2</task-retry-limit>
+        </retry-parameters>
+    </queue>
+
+    <queue>
+        <name>admin-prepare-email-group-mode-queue</name>
         <rate>1/s</rate>
         <bucket-size>1</bucket-size>
         <retry-parameters>
